Make `default-board` accept partial board information
`arduino-cli--default-board` now handles partial board information,
the primary use of which is the ability to compile with only
arduino-cli-default-fqbn specified. Appropriate error checks have been
added to make the use of this feature clear, such as the package
giving you appropriate feedback when trying to `compile-and-upload`
with missing port/fqbn information.

Note that `arduino-cli-default-fqbn` and `arduino-cli-default-port`
have been renamed to drop the cli-mode midfix

fix #1