Make declared types of user options match their documentation.
This affects:

 * arduino-cli-default-fqbn: originally string, allow nil too

 * arduino-cli-default-port: originally string, allow nil too

 * arduino-cli-warnings: originally boolean, but instead of t the
   documented non-nil values are 'default, 'more, or 'all

 * arduino-cli-verbosity: originally boolean, but instead of t the
   documented non-nil values are 'quiet or 'verbose

This change prevents the warnings that were generated when setting
these customizable variables to values that didn't match the declared
type (which is what alerted me to the difference).  Also, having declared
types that encompass all documented values means that the Easy
Customization Interface can be used to edit these options.

The tests check that each of user option can be assigned all documented
values without generating a warning.

The simple makefile provides a way to run all the tests with M-x
compile.  The makefile is structured to accommodate multiple files of
tests, though I have only included one in this commit.